PDS-656 - Added file containing registered context products (#3)
* PDS-656 - Added file containing registered context products

This commit includes the following:
* A python script that queries the EN Search Service by default for
Context Products and creates a JSON file containing a list of LIDVIDs
found.
* A JSON file that represents the list of Context Products that are
found in the EN Search Service at the time the script was run. This will
be used to check that context product references in a label exist in
this list. A follow on ticket will be created to do the actual
implementation.

The registered_context_products.json will reside in a resources/ folder
in the Validate Tool release package. Install documentation was updated
describing this new file.