PDS-678 - Add Context Product reference validation (#6)
* PDS-678 - Add Context Product reference validation

Added the capability to check that Context Products referenced in a
product label exist within a supplied list of Context Products
registered at the Engineering Node. If a reference is determined to not
exist within the supplied list of registered references, an error of
type "error.label.context_ref_not_found" is thrown to the report.

* Fixed PARENT_DIR setting in Windows batch file.

* Fixed issue that was throwing exceptions in the report

While Ed was testing the Context Product Reference validation feature,
he noticed that the tool was throwing exceptions when validating using
Bundle rules. Issue was that the new feature settings were not being
propagated down to the child contexts.